fix nxtHttp status bug

+++ bitway2/src/main/scala/com/coinport/bitway/NxtBitway/model/NxtModel.scala
@@ -14,7 +14,7 @@ case class NxtAddress(
 case class NxtBlockStatus(
                            lastBlockHeight: Long,
                            lastBlockId: String,
-                           timestamp: Long)
+                           timestamp: Double)
 
 case class NxtTransaction(
                            transactionId: String,
